# backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Verified B2B Deal Network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{orgId}/verification-cases:
    get:
      operationId: listVerificationCases
      tags: [verification]
      summary: List verification cases for an organization
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VerificationCase"
    post:
      operationId: submitVerificationCase
      tags: [verification]
      summary: Submit a new verification case
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Case submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationCase"
  /organizations/{orgId}/attestation:
    get:
      operationId: getVerificationAttestation
      tags: [verification]
      summary: Get current verification attestation
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Attestation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationAttestation"
        "404":
          description: No attestation found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Documents ───────────────────────────────────────────────────────────────
...
  /organizations/{orgId}/wallet:
    get:
      operationId: getWallet
      tags: [wallets]
      summary: Get organization wallet info and balance
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletInfo"
  /organizations/{orgId}/wallet/transfer:
    post:
      operationId: initiateTransfer
      tags: [wallets]
      summary: Initiate a transfer from the organization wallet
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferInput"
      responses:
        "202":
          description: Transfer queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"

  # ─── Admin ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ────────────────────────────────────────────────────────────────────
...
    VerificationCase:
      type: object
      additionalProperties: false
      required: [id, organizationId, status, createdAt]
      properties:
        id:
          type: string
        organizationId:
          type: string
        status:
          type: string
          enum: [SUBMITTED, IN_REVIEW, APPROVED, REJECTED, MORE_INFO_NEEDED]
        reviewerAdminId:
          type: ["string", "null"]
        decisionNotes:
          type: ["string", "null"]
        decidedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    VerificationAttestation:
      type: object
      additionalProperties: false
      required: [id, organizationId, claim, issuer, issuedAt, expiresAt, revoked]
      properties:
        id:
          type: string
        organizationId:
          type: string
        claim:
          type: string
        issuer:
          type: string
        reviewerId:
          type: ["string", "null"]
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        revoked:
          type: boolean

    # ── Documents ───────────────────────────────────────────────────────────────
...
    WalletInfo:
      type: object
      additionalProperties: false
      required: [id, organizationId, circleWalletId, blockchainAddress, chain, walletType, purpose, createdAt]
      properties:
        id:
          type: string
        organizationId:
          type: string
        circleWalletId:
          type: string
        blockchainAddress:
          type: string
        chain:
          type: string
        walletType:
          type: string
          enum: [USER_CONTROLLED, DEVELOPER_CONTROLLED]
        purpose:
          type: string
          enum: [TREASURY, PROCUREMENT, SALES_RECEIPT, PLATFORM]
        balances:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              amount:
                type: string
              currency:
                type: string
        createdAt:
          type: string
          format: date-time
    TransferInput:
      type: object
      additionalProperties: false
      required: [amount, currency, destinationAddress, destinationChain]
      properties:
        amount:
          type: string
        currency:
          type: string
        destinationAddress:
          type: string
        destinationChain:
          type: string
        idempotencyKey:
          type: string
    TransferResponse:
      type: object
      additionalProperties: false
      required: [transferId, status]
      properties:
        transferId:
          type: string
        status:
          type: string

    # ── Connections ──────────────────────────────────────────────────────────────
...